pyproject.toml
[project]
name = "academic-assistant"
version = "0.1.0"
description = "Academic assistant with PDF parsing, semantic chunking, RAG retrieval, and LLM answer generation"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "sqlalchemy==2.0.43",
    "pyjwt==2.10.1",
    "python-multipart==0.0.20",
    "ollama==0.3.3",
    "pdfplumber==0.11.4",
    "faiss-cpu==1.10.0",
    "numpy==1.26.4",
    "sentence-transformers==2.7.0",
    "google-generativeai==0.7.2",
    "python-dotenv>=1.2.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.3",
    "google-auth-oauthlib==1.2.2",
    "cryptography==43.0.1",
    "tqdm==4.66.4",
    "tzdata>=2026.1",
]
